% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFScore.R
\name{AFScore}
\alias{AFScore}
\title{Calculate the correlation activity fluctuation score}
\usage{
AFScore(MEs_table, comp_group, scale = FALSE)
}
\arguments{
\item{MEs_table}{A data frame contains the eigengene values of each module, obtained in \code{\link{Mod_Extra}}}

\item{comp_group}{The vector contains cell group/type class identity. The length of the vector must be equal to the number of rows of MEs_table}

\item{scale}{Logical. If TRUE, the score will be scaled according to group}
}
\value{
a dataframe of AFscores for each cell group/type
}
\description{
This function can assess the degree to which the module's activity fluctuates
}
\details{
The order of the comp_group, and rownames of the MEs_table should correspond to each other
}
