% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCScore.R
\name{CCScore}
\alias{CCScore}
\title{Calculate the correlation coefficient score}
\usage{
CCScore(MEs_table, comp_group, values, scale = FALSE)
}
\arguments{
\item{MEs_table}{A data.frame contains the eigengene values of each module, obtained in \code{\link{Mod_Extra}}}

\item{comp_group}{The vector contains cell group/type class identity. The length of the vector must be equal to the number of rows of MEs_table}

\item{values}{The trait values with which to compare, corresponding with comp_group. The value must be numerical and the length of which must be equal to the length of comp_group}

\item{scale}{Logical. If TRUE, the score will be scaled according to group.}
}
\value{
a dataframe of CCscores for each cell group/type
}
\description{
Assessment of the relationship between the trait of cell type and the module activity
}
\details{
The order of the comp_group, values and rownames of the MEs_table should correspond to each other
}
