% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter data}
\usage{
filter_data(data_Expr, mad = TRUE, mad.thre = NULL, goodsamples = TRUE)
}
\arguments{
\item{data_Expr}{A data.frame generated from \code{\link{Mean_Expre}}}

\item{mad}{Logical. Whether to use the MAD values to filter data. Default is TRUE}

\item{mad.thre}{The threshold to filter the mad.Default is max(max(m.mad)*0.25,0.01)}

\item{goodsamples}{Logical. Whether to use the goodSamplesGenes function to filter data. Default is TRUE}
}
\value{
a data.frame with filtered values
}
\description{
filter the data with Mean Absolute Deviation and goodsample function
}
\details{
MAD values are used to screen for features with variability, and the goodSamplesGenes function is used to check for missing values
}
