% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ide_Hub.R, R/Mean_Expre.R
\name{Ide_Hub}
\alias{Ide_Hub}
\alias{Mean_Expre}
\title{Hub genes identification}
\usage{
Ide_Hub(net, data_Expr, KeyMod, markers = NULL, mod = NULL, sftpower = NULL)

Mean_Expre(
  object,
  group.1,
  group.2 = NULL,
  group.3 = NULL,
  features = NULL,
  slot = "count"
)
}
\arguments{
\item{net}{A list obtained in \code{\link{Mod_Infer}}}

\item{data_Expr}{A data.frame generated from \code{\link{Mean_Expre}}}

\item{KeyMod}{The dataframe of identified key modules obtained in \code{\link{Ide_KeyMod}}}

\item{markers}{A dataframe of marker genes of each given cell groups}

\item{mod}{A list generated from \code{\link{Mod_Extra}}}

\item{sftpower}{The soft-threshold inferred from Mod_Infer}

\item{object}{A Seurat object.}

\item{group.1}{The group information provided by users, which must be an identity class of the Seurat object}

\item{group.2}{The second group information provided by users, which must be an identity class of the Seurat object}

\item{group.3}{The third group information provided by users, which must be an identity class of the Seurat object}

\item{features}{Genes to used.Default is to use all genes}

\item{slot}{Slot to pull data from..Default slot will be "counts"}
}
\value{
a dataframe of hub genes for each key module of each cell group/type

a data.frame of mean expression of features in each cell states
}
\description{
This function is able to identify hub genes for key modules

Process the Seurat Object. According to the group information provided by users, cells were divided into cell states and the average expression value of features(all the features of Seurat object or provided by the user) in each cell state was obtained for subsequent gene module inference
}
\details{
If markers is NULL, we will calculate the logFC values between the query cell type and others;
         If mod id NULL, this function will generated a list containing members and ME values of each module

The order of groups is interchangeable, but group.3 cannot be set if group.2 is Null. The as_matrix function was used to transform a sparse matrix into an ordinary one in a faster and more memory efficient manner
}
