% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ide_KeyMod.R
\name{Ide_KeyMod}
\alias{Ide_KeyMod}
\title{Key modules identification}
\usage{
Ide_KeyMod(
  MEs_table,
  MS_score = NULL,
  comp_group = NULL,
  values = NULL,
  Scc_scale = FALSE,
  Saf_scale = FALSE
)
}
\arguments{
\item{MEs_table}{A data.frame contains the eigengene values of each module, obtained in \code{\link{Mod_Extra}}}

\item{MS_score}{A data.frame contains the module significance score from \code{\link{MSScore}}. If NULL, the score will be calculated}

\item{comp_group}{The vector contains cell group/type class identity. The length of the vector must be equal to the number of rows of MEs_table}

\item{values}{The trait values with which to compare, corresponding with comp_group. The value must be numerical and the length of which must be equal to the length of comp_group}

\item{Scc_scale}{Logical. If TRUE, Scaling is performed when calculating the correlation coefficient score}

\item{Saf_scale}{Logical. If TRUE, Scaling is performed when calculating the activity fluctuation score}
}
\value{
a dataframe of key modules for each given group
}
\description{
This function is able to identify key modules for given cell groups according to the module significance score
}
\details{
This function will also export the absolute values of module significance score of corresponding key module
}
