% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mod_Infer.R
\name{Mod_Infer}
\alias{Mod_Infer}
\title{WGCNA module inference}
\usage{
Mod_Infer(
  data_Expr,
  sftpower = NULL,
  RsquaredCut = 0.9,
  maxBlockSize = "total",
  deepSplit = 4,
  minModuleSize = 20,
  mergeCutHeight = 0.1
)
}
\arguments{
\item{data_Expr}{A data.frame generated from \code{\link{Mean_Expre}}}

\item{sftpower}{The soft-threshold to identify scale-free gene network}

\item{RsquaredCut}{The desired minimum scale free topology fitting index. Default is 0.9}

\item{maxBlockSize}{Maximum cluster gap given as the fraction.Default is all of genes of given dataset}

\item{deepSplit}{Parameter that control the resolution sensitivity of the module}

\item{minModuleSize}{Min cluster gap given as the fraction.Default is 20}

\item{mergeCutHeight}{Parameter for module merging.The larger this parameter is, the fewer modules there are}
}
\value{
A list of inferred network information. Details can be seen in \code{\link[WGCNA]{blockwiseModules}}
}
\description{
According to the input expression matrix, the WGCNA gene module was obtained
}
\details{
When sftpower is NULL, the appropriate soft threshold is inferred through scale-free network analysis. And the inferred or input sftpower would be in the name of 'sftpower' of the exported list
}
