% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSScore.R
\name{MSScore}
\alias{MSScore}
\title{Calculate the module significance score}
\usage{
MSScore(MEs_table, comp_group, values, Scc_scale = FALSE, Saf_scale = FALSE)
}
\arguments{
\item{MEs_table}{A data.frame contains the eigengene values of each module, obtained in \code{\link{Mod_Extra}}.}

\item{comp_group}{The vector contains cell group/type class identity. The length of the vector must be equal to the number of rows of MEs_table}

\item{values}{The trait values with which to compare, corresponding with comp_group. The value must be numerical and the length of which must be equal to the length of comp_group}

\item{Scc_scale}{Logical. If TRUE, Scaling is performed when calculating the correlation coefficient score}

\item{Saf_scale}{Logical. If TRUE, Scaling is performed when calculating the activity fluctuation score}
}
\value{
a dataframe of MSscores for each cell group/type
}
\description{
For each cell group/type, this function can assess the comprehensive correlation of inferred gene modules by scoring them
}
\details{
The correlation coefficient score and activity fluctuation score will be calculated together with module significance score with MSScore function. If you want to obtain them respectively, run \code{\link{CCScore}} and \code{\link{AFScore}} seperately
         The order of the comp_group, values and rownames of the MEs_table should correspond to each other
}
